+++ swap/src/cli/command.rs
@@ -1,5 +1,6 @@
 use crate::env::GetConfig;
 use crate::fs::system_data_dir;
+use crate::network::rendezvous::{XmrBtcNamespace, DEFAULT_RENDEZVOUS_ADDRESS};
 use crate::{env, monero};
 use anyhow::{Context, Result};
 use libp2p::core::Multiaddr;
@@ -188,6 +189,20 @@ where
                 bitcoin_target_block: bitcoin_target_block_from(bitcoin_target_block, is_testnet),
             },
         },
+        RawCommand::ListSellers {
+            rendezvous_node_addr,
+            tor: Tor { tor_socks5_port },
+        } => Arguments {
+            env_config: env_config_from(is_testnet),
+            debug,
+            json,
+            data_dir: data::data_dir_from(data, is_testnet)?,
+            cmd: Command::ListSellers {
+                rendezvous_node_addr,
+                namespace: rendezvous_namespace_from(is_testnet),
+                tor_socks5_port,
+            },
+        },
     };
 
     Ok(ParseResult::Arguments(arguments))
@@ -223,6 +238,11 @@ pub enum Command {
         force: bool,
         bitcoin_electrum_rpc_url: Url,
         bitcoin_target_block: usize,
+    },
+    ListSellers {
+        rendezvous_node_addr: Multiaddr,
+        namespace: XmrBtcNamespace,
+        tor_socks5_port: u16,
     },
 }
 
@@ -310,6 +330,17 @@ pub enum RawCommand {
 
         #[structopt(flatten)]
         bitcoin: Bitcoin,
+    },
+    ListSellers {
+        #[structopt(
+            long,
+            help = "The multiaddr (including peer-id) of a rendezvous node that sellers register with",
+            default_value = DEFAULT_RENDEZVOUS_ADDRESS
+        )]
+        rendezvous_node_addr: Multiaddr,
+
+        #[structopt(flatten)]
+        tor: Tor,
     },
 }
 
@@ -403,6 +434,14 @@ fn bitcoin_electrum_rpc_url_from(url: Option<Url>, testnet: bool) -> Result<Url>
         Ok(Url::from_str(DEFAULT_ELECTRUM_RPC_URL_TESTNET)?)
     } else {
         Ok(Url::from_str(DEFAULT_ELECTRUM_RPC_URL)?)
+    }
+}
+
+fn rendezvous_namespace_from(is_testnet: bool) -> XmrBtcNamespace {
+    if is_testnet {
+        XmrBtcNamespace::Testnet
+    } else {
+        XmrBtcNamespace::Mainnet
     }
 }