+++ swap/tests/harness/mod.rs
@@ -236,6 +236,7 @@ async fn start_alice(
         latest_rate,
         resume_only,
         env_config,
+        None,
     )
     .unwrap();
     swarm.listen_on(listen_address).unwrap();
@@ -445,12 +446,16 @@ impl BobParams {
     ) -> Result<(cli::EventLoop, cli::EventLoopHandle)> {
         let tor_socks5_port = get_port()
             .expect("We don't care about Tor in the tests so we get a free port to disable it.");
-        let mut swarm = swarm::cli(
-            &self.seed,
+
+        let behaviour = cli::Behaviour::new(
             self.alice_peer_id,
-            tor_socks5_port,
             self.env_config,
             self.bitcoin_wallet.clone(),
+        );
+        let mut swarm = swarm::cli(
+            self.seed.derive_libp2p_identity(),
+            tor_socks5_port,
+            behaviour,
         )
         .await?;
         swarm